FPVTL-842: Change naming to solve test
