+++ src/test/java/uk/gov/hmcts/reform/prl/controllers/caseassignment/CaseBarristerAssignmentControllerTest.java
@@ -1,6 +1,5 @@
-package uk.gov.hmcts.reform.prl.controllers;
+package uk.gov.hmcts.reform.prl.controllers.caseassignment;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -15,6 +14,7 @@ import uk.gov.hmcts.reform.ccd.client.model.CallbackRequest;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 import uk.gov.hmcts.reform.prl.clients.ccd.CaseAssignmentService;
 import uk.gov.hmcts.reform.prl.enums.noticeofchange.BarristerRole;
+import uk.gov.hmcts.reform.prl.exception.GrantCaseAccessException;
 import uk.gov.hmcts.reform.prl.models.Organisation;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicList;
 import uk.gov.hmcts.reform.prl.models.common.dynamic.DynamicListElement;
@@ -42,6 +42,7 @@ import static org.mockito.ArgumentMatchers.anyMap;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.doAnswer;
+import static org.mockito.Mockito.doThrow;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
@@ -49,7 +50,7 @@ import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.ALLOCATED_BARRI
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.INVALID_CLIENT;
 
 @ExtendWith(MockitoExtension.class)
-class CaseAssignmentControllerTest {
+class CaseBarristerAssignmentControllerTest {
     @Mock
     private CaseAssignmentService caseAssignmentService;
     @Mock
@@ -101,8 +102,8 @@ class CaseAssignmentControllerTest {
     }
 
     @Test
-    void testSuccessSubmitAddBarrister() throws JsonProcessingException {
-        System.out.println(objectMapper.writerWithDefaultPrettyPrinter().writeValueAsString(allocatedBarrister));
+    void testSuccessSubmitAddBarrister() {
+
         Optional<String> userId = Optional.of("userId");
         when(authorisationService.isAuthorized(any(), any()))
             .thenReturn(true);
@@ -133,6 +134,9 @@ class CaseAssignmentControllerTest {
             callbackRequest
         );
 
+        assertThat(response.getData().get(ALLOCATED_BARRISTER))
+                       .isNull();
+
         assertThat(response.getErrors()).isEmpty();
 
         verify(caseAssignmentService).validateAddRequest(eq(userId),
@@ -145,6 +149,61 @@ class CaseAssignmentControllerTest {
                                                    eq(barristerRole.get()),
                                                    isA(AllocatedBarrister.class));
     }
+
+    @Test
+    void testGrantCaseAccessExceptionOnSubmitAddBarrister() {
+
+        Optional<String> userId = Optional.of("userId");
+        when(authorisationService.isAuthorized(any(), any()))
+            .thenReturn(true);
+        when(organisationService.findUserByEmail(allocatedBarrister.getBarristerEmail()))
+            .thenReturn(userId);
+        Optional<String> barristerRole = Optional.of(BarristerRole.C100APPLICANTBARRISTER1.getCaseRoleLabel());
+
+        when(caseAssignmentService.deriveBarristerRole(anyMap(), isA(CaseData.class), isA(AllocatedBarrister.class)))
+            .thenReturn(barristerRole);
+
+        doThrow(new GrantCaseAccessException("User(s) not granted [C100APPLICANTBARRISTER3] to the case "))
+            .when(caseAssignmentService).addBarrister(isA(CaseData.class),
+                                                     eq(userId.get()),
+                                                     eq(barristerRole.get()),
+                                                     isA(AllocatedBarrister.class));
+
+
+        Map<String, Object> caseData = new HashMap<>();
+        caseData.put(ALLOCATED_BARRISTER, allocatedBarrister);
+
+        CaseDetails caseDetails = CaseDetails.builder()
+            .id(1234L)
+            .createdDate(LocalDateTime.now())
+            .lastModified(LocalDateTime.now())
+            .data(caseData)
+            .build();
+
+        CallbackRequest callbackRequest = CallbackRequest.builder()
+            .caseDetails(caseDetails)
+            .build();
+
+        AboutToStartOrSubmitCallbackResponse response = caseAssignmentController.submitAddBarrister(
+            "auth",
+            "s2sToken",
+            callbackRequest
+        );
+
+        assertThat(response.getErrors())
+            .contains("User(s) not granted [C100APPLICANTBARRISTER3] to the case ");
+
+        verify(caseAssignmentService).validateAddRequest(eq(userId),
+                                                         isA(CaseData.class),
+                                                         eq(barristerRole),
+                                                         isA(AllocatedBarrister.class),
+                                                         anyList());
+        verify(caseAssignmentService).addBarrister(isA(CaseData.class),
+                                                   eq(userId.get()),
+                                                   eq(barristerRole.get()),
+                                                   isA(AllocatedBarrister.class));
+    }
+
 
     @Test
     void testErrorsSubmitAddBarrister() {
@@ -294,6 +353,9 @@ class CaseAssignmentControllerTest {
             callbackRequest
         );
 
+        assertThat(response.getData().get(ALLOCATED_BARRISTER))
+            .isNull();
+
         assertThat(response.getErrors()).isEmpty();
 
         String selectedPartyId = allocatedBarrister.getPartyList().getValueCode();
@@ -322,12 +384,13 @@ class CaseAssignmentControllerTest {
             .caseDetails(caseDetails)
             .build();
 
+        String selectedPartyId = allocatedBarrister.getPartyList().getValueCode();
         doAnswer(invocation -> {
             List<String> errors = invocation.getArgument(2);
             errors.add("errors");
             return null;
         }).when(caseAssignmentService).validateRemoveRequest(isA(CaseData.class),
-                                                             eq(allocatedBarrister.getPartyList().getValueCode()),
+                                                             eq(selectedPartyId),
                                                              anyList());
 
         AboutToStartOrSubmitCallbackResponse response = caseAssignmentController.submitRemoveBarrister(
@@ -339,10 +402,10 @@ class CaseAssignmentControllerTest {
         assertThat(response.getErrors()).contains("errors");
 
         verify(caseAssignmentService).validateRemoveRequest(isA(CaseData.class),
-                                                         eq(allocatedBarrister.getPartyList().getValueCode()),
+                                                         eq(selectedPartyId),
                                                          anyList());
         verify(caseAssignmentService, never()).removeBarrister(isA(CaseData.class),
-                                                   eq(allocatedBarrister.getPartyList().getValueCode()));
+                                                   eq(selectedPartyId));
     }
 
     @Test
